requirements.txt
numpy==1.24.1
pandas==1.5.2
scipy==1.9.0
flask==2.0.1
Werkzeug==2.2.2
paddlepaddle==2.4.2
paddlenlp==2.5.2
trustai >= 0.1.7
redis==3.5.3
#rabbitmq
pika==1.1.0
pymysql==0.10.1
dbutils==3.0.0
celery==5.2.1

scikit-learn~=1.2.1
jieba~=0.42.1
ipython~=8.8.0